Adds types and routines for creating CUDA Arrays
The primary new type is ArrayObject, which represents a generic array
with up to 3-dimensions.

An ArrayObject has a set of dimensions, an ArrayFormat, a number of
channels (1, 2, or 4), and a set of ArrayObjectFlags

Adds tests that assert correct ArrayObject values are accepted and that
bad ones are rejected with a panic.